requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
pandas>=1.4.0
scikit-learn>=1.0.0
pillow>=9.0.0
numpy>=1.22.0
matplotlib>=3.5.0
opencv-python>=4.5.5
albumentations>=1.1.0
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.8.0